+++ examples/drop_down_table/drop_down_grouped/index.jsx
@@ -1,5 +1,5 @@
 import { useRef, useState } from 'react';
-import { DropDownTable, useLayoutListBox, layoutMaxWidth, layoutMaxHeight } from '../../../src';
+import { DropDownTable, layoutMaxWidth, layoutMaxHeight } from '../../../src';
 import cats from '../../data/cats.json';
 
 const columns = ['breed', 'country', 'origin', 'bodyType', 'pattern'];
@@ -8,10 +8,11 @@ function mapOption({ breed, coatLength }) {
   return { label: breed, group: coatLength };
 }
 
+const onLayoutListBox = [layoutMaxWidth, layoutMaxHeight];
+
 export function Example() {
   const [value, setValue] = useState(null);
   const ref = useRef();
-  const onLayoutListBox = useLayoutListBox(layoutMaxWidth, layoutMaxHeight);
   const [managedFocus, setManagedFocus] = useState(true);
 
   return (