+++ src/hooks/index.js
@@ -1,4 +1,3 @@
 export { usePrefixSearch } from './use_prefix_search';
 export { useAsyncSearch } from './use_async_search';
 export { useTokenSearch } from './use_token_search';
-export { useLayoutListBox } from './use_layout_list_box';